+++ engines/store/mongo/structs/word.go
@@ -8,7 +8,7 @@ type Word struct {
 	BoardID      objectid.ObjectID `bson:"board_id"`
 	Text         string            `bson:"text"`
 	Translations []string          `bson:"translations"`
-	Weight       int64             `bson:"weight"`
+	Weight       float64           `bson:"weight"`
 	ShownTimes   int64             `bson:"shown_times"`
 	Fails        int64             `bson:"fails"`
 	LastAnswered int64             `bson:"last_answered_unix_nano"`
